moved types to separate file, refactor
