+++ src/containers/Application/Application.tsx
@@ -1,8 +1,7 @@
 import { useState } from "react";
-import { FormActions } from "../../FormBuilder/formUtils";
 
 import ApplicationForm from "../../forms/ApplicationForm/ApplicationForm";
-import { FormValues } from "../../FormBuilder/formUtils";
+import { FormValues, FormActions } from "../../FormBuilder";
 
 /**
  * Component to render form and ancillary elements related to the form.
